backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
psycopg[binary]==3.1.18
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1
python-dotenv==1.0.1
email-validator==2.1.1
google-generativeai==0.5.4
requests==2.31.0
Pillow==10.3.0
openai==1.12.0
numpy==1.26.4
gradio_client==1.0.1
gunicorn
